fix: consolidated resample into cog
