+++ packages/cog/src/__test__/gdal.test.ts
@@ -1,5 +1,5 @@
-import { GdalCogBuilder } from '../gdal';
 import * as o from 'ospec';
+import { GdalCogBuilder } from '../gdal';
 
 o.spec('GdalCogBuilder', () => {
     o('should default all options', () => {
@@ -7,7 +7,7 @@ o.spec('GdalCogBuilder', () => {
 
         o(builder.config.bbox).equals(undefined);
         o(builder.config.compression).equals('webp');
-        o(builder.config.resampling).equals('lanczos');
+        o(builder.config.resampling).equals('cubic');
         o(builder.config.blockSize).equals(512);
         o(builder.config.alignmentLevels).equals(1);
     });
