+++ packages/cog/src/cli/actions/action.cog.ts
@@ -1,21 +1,17 @@
 import { Env, FileOperator, FileOperatorSimple, LogConfig, LogType } from '@basemaps/lambda-shared';
-import {
-    CommandLineAction,
-    CommandLineFlagParameter,
-    CommandLineIntegerParameter,
-    CommandLineStringParameter,
-} from '@microsoft/ts-command-line';
+import { CommandLineAction, CommandLineFlagParameter, CommandLineIntegerParameter, CommandLineStringParameter } from '@microsoft/ts-command-line';
 import { createReadStream, promises as fs } from 'fs';
 import * as ulid from 'ulid';
 import { buildCogForQuadKey, CogJob } from '../../cog';
 import { buildWarpedVrt } from '../../cog.vrt';
-import { makeTempFolder, getJobPath } from '../folder';
+import { getJobPath, makeTempFolder } from '../folder';
 
 export class ActionCogCreate extends CommandLineAction {
     private job?: CommandLineStringParameter;
     private quadKey?: CommandLineStringParameter;
     private commit?: CommandLineFlagParameter;
     private quadKeyIndex?: CommandLineIntegerParameter;
+    private resampleMethod?: CommandLineStringParameter;
 
     public constructor() {
         super({
@@ -69,6 +65,8 @@ export class ActionCogCreate extends CommandLineAction {
 
         const isCommit = this.commit?.value ?? false;
 
+        const resampleMethod = this.resampleMethod?.value ?? 'cubic';
+
         const logger = LogConfig.get().child({ id: processId, correlationId: job.id, imageryName: job.name });
         LogConfig.set(logger);
 
@@ -98,7 +96,7 @@ export class ActionCogCreate extends CommandLineAction {
             // Sometimes we need to force a epsg3857 projection to get the COG to build since its fast just do it locally
             const vrtPath = await buildWarpedVrt(job, tmpVrt, job.output.vrt.options, tmpFolder, logger);
 
-            await buildCogForQuadKey(job, quadKey, vrtPath, tmpTiff, logger, isCommit);
+            await buildCogForQuadKey(job, quadKey, vrtPath, tmpTiff, logger, isCommit, resampleMethod);
             logger.info({ target: targetPath }, 'StoreTiff');
             if (isCommit) {
                 await outputFs.write(targetPath, createReadStream(tmpTiff), logger);
@@ -136,6 +134,13 @@ export class ActionCogCreate extends CommandLineAction {
         this.commit = this.defineFlagParameter({
             parameterLongName: '--commit',
             description: 'Begin the transformation',
+            required: false,
+        });
+
+        this.resampleMethod = this.defineStringParameter({
+            argumentName: 'RESAMPLE',
+            parameterLongName: '--resample-method',
+            description: 'resampling method to use',
             required: false,
         });
     }
