+++ packages/cog/src/gdal.ts
@@ -56,7 +56,7 @@ export class GdalCogBuilder {
             bbox: config.bbox,
             alignmentLevels: config.alignmentLevels ?? 1,
             compression: config.compression ?? 'webp',
-            resampling: config.resampling ?? 'lanczos',
+            resampling: config.resampling ?? 'cubic',
             blockSize: config.blockSize ?? 512,
         };
